requirements.txt
google-generativeai==0.8.4
gradio==5.0.0
Flask==3.1.0